requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.4.4
psycopg-binary==3.1.18
django-phonenumber-field==7.3.0
phonenumbers==8.13.29